requirements.txt
PySide6
pydivert
psutil
pywin32
pandas
matplotlib
requests
dnspython
cryptography
schedule
